move `decode_engine_core_outputs`
